pyproject.toml
[project]
name = "pdf-assistant-api"
version = "0.1.0"
description = ""
authors = [
    { name = "None" }
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "aiohappyeyeballs (==2.4.4)",
    "aiohttp (==3.11.11)",
    "aiosignal (==1.3.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.7.0)",
    "attrs (==24.3.0)",
    "auth0-python (==4.7.2)",
    "blinker (==1.9.0)",
    "cachetools (==5.5.0)",
    "certifi (==2024.12.14)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.0)",
    "click (==8.1.7)",
    "cryptography (==43.0.3)",
    "distro (==1.9.0)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.6)",
    "fastapi-cli (==0.0.7)",
    "frozenlist (==1.5.0)",
    "google-api-core (==2.24.0)",
    "google-auth (==2.37.0)",
    "google-auth-stubs (==0.3.0)",
    "google-cloud-core (==2.4.1)",
    "google-cloud-secret-manager (==2.22.0)",
    "google-cloud-storage (==2.19.0)",
    "google-cloud-tasks (==2.18.0)",
    "google-crc32c (==1.6.0)",
    "google-resumable-media (==2.7.2)",
    "googleapis-common-protos (==1.66.0)",
    "grpc-google-iam-v1 (==0.13.1)",
    "grpc-stubs (==1.53.0.5)",
    "grpcio (==1.68.1)",
    "grpcio-status (==1.68.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.0.0)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.4)",
    "jiter (==0.8.2)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "multidict (==6.1.0)",
    "mypy (==1.13.0)",
    "mypy-extensions (==1.0.0)",
    "openai (==1.58.1)",
    "orjson (==3.10.12)",
    "packaging (==24.2)",
    "pluggy (==1.5.0)",
    "propcache (==0.2.1)",
    "proto-plus (==1.25.0)",
    "protobuf (==5.29.2)",
    "psycopg2 (==2.9.10)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.1)",
    "pycparser (==2.22)",
    "pydantic (==2.10.3)",
    "pydantic-extra-types (==2.10.1)",
    "pydantic-settings (==2.7.0)",
    "pydantic-core (==2.27.1)",
    "pygments (==2.18.0)",
    "pyjwt (==2.10.1)",
    "pymysql (==1.1.1)",
    "pytest (==8.3.4)",
    "python-dotenv (==1.0.1)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.3)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.12.0)",
    "rsa (==4.9)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.41.3)",
    "tqdm (==4.67.1)",
    "typer (==0.15.1)",
    "types-requests (==2.32.0.20241016)",
    "typing-extensions (==4.12.2)",
    "ujson (==5.10.0)",
    "urllib3 (==2.2.3)",
    "uvicorn (==0.34.0)",
    "uvloop (==0.21.0)",
    "watchfiles (==1.0.3)",
    "websockets (==14.1)",
    "werkzeug (==3.1.3)",
    "yarl (==1.18.3)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "google-cloud-firestore (>=2.20.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[tool.mypy]
python_version = "3.13"
strict = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
exclude = "(^|/|\\\\)venv($|/|\\\\)"
[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "auth0.*"
ignore_missing_imports = true